ch5 security

+++ ch4-spring-db2/ch4-labs/src/main/java/com/example/ch4labs/dto/Comment/CommentCreateRequest.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class CommentCreateRequest {
     private String content;
     private String author;
-    private Long parentCommentId;
+    private Integer parentCommentId;
 }
